CrazyLogin: added registered Filter to list command, added account create command, code improvements, refactoring

+++ CrazyAPI_Login/src/de/st_ddt/crazylogin/LoginPlugin.java
@@ -75,7 +75,7 @@ public interface LoginPlugin extends CrazyPluginInterface
 
 	public abstract int getMaxRegistrationsPerIP();
 
-	public abstract int getMoveRange();
+	public abstract double getMoveRange();
 
 	public abstract int getMinNameLength();
